// src/shared/api/orval/types/noticeCategory.ts
/**
 * Generated by orval v7.17.0 🍺
 * Do not edit manually.
 * Smarter Store API
 * Smarter Store 백엔드 API 문서

**관리자 대시보드**: [매출 현황 대시보드](/admin/dashboard.html)
 * OpenAPI spec version: 1.0.0
 */

/**
 * 
        공지사항 카테고리
        - BOOKING_NOTICE: 예매 유의사항
        - BANK_TRANSFER_NOTICE: 무통장입금 입금 시 주의사항
        - TICKET_RECEIPT_GUIDE: 티켓 수령안내
        - MOBILE_TICKET_GUIDE: 모바일 티켓 안내
        - REFUND_GUIDE: 환불 안내
        - CANCELLATION_REFUND_NOTICE: 취소 및 환불 유의사항
    
 */
export type NoticeCategory = typeof NoticeCategory[keyof typeof NoticeCategory];


// eslint-disable-next-line @typescript-eslint/no-redeclare
export const NoticeCategory = {
  BOOKING_NOTICE: 'BOOKING_NOTICE',
  BANK_TRANSFER_NOTICE: 'BANK_TRANSFER_NOTICE',
  TICKET_RECEIPT_GUIDE: 'TICKET_RECEIPT_GUIDE',
  MOBILE_TICKET_GUIDE: 'MOBILE_TICKET_GUIDE',
  REFUND_GUIDE: 'REFUND_GUIDE',
  CANCELLATION_REFUND_NOTICE: 'CANCELLATION_REFUND_NOTICE',
} as const;